+++ examples/notification/main.tf
@@ -32,43 +32,42 @@ resource "null_resource" "download_package" {
   }
 }
 
-data "null_data_source" "downloaded_package" {
-  inputs = {
-    id       = null_resource.download_package.id
-    filename = local.downloaded
-  }
-}
-
 module "lambda_function1" {
   source  = "terraform-aws-modules/lambda/aws"
-  version = "~> 1.0"
+  version = "~> 2.0"
 
   function_name = "${random_pet.this.id}-lambda1"
   handler       = "index.lambda_handler"
   runtime       = "python3.8"
 
   create_package         = false
-  local_existing_package = data.null_data_source.downloaded_package.outputs["filename"]
+  local_existing_package = local.downloaded
 }
 
 module "lambda_function2" {
   source  = "terraform-aws-modules/lambda/aws"
-  version = "~> 1.0"
+  version = "~> 2.0"
 
   function_name = "${random_pet.this.id}-lambda2"
   handler       = "index.lambda_handler"
   runtime       = "python3.8"
 
   create_package         = false
-  local_existing_package = data.null_data_source.downloaded_package.outputs["filename"]
+  local_existing_package = local.downloaded
 }
 
 module "sns_topic1" {
-  source = "terraform-aws-modules/cloudwatch/aws//examples/fixtures/aws_sns_topic"
+  source  = "terraform-aws-modules/sns/aws"
+  version = "~> 3.0"
+
+  name_prefix = "${random_pet.this.id}-2"
 }
 
 module "sns_topic2" {
-  source = "terraform-aws-modules/cloudwatch/aws//examples/fixtures/aws_sns_topic"
+  source  = "terraform-aws-modules/sns/aws"
+  version = "~> 3.0"
+
+  name_prefix = "${random_pet.this.id}-2"
 }
 
 resource "aws_sqs_queue" "this" {
@@ -99,22 +98,22 @@ resource "aws_sqs_queue_policy" "allow_external" {
 module "all_notifications" {
   source = "../../modules/notification"
 
-  bucket = module.s3_bucket.this_s3_bucket_id
+  bucket = module.s3_bucket.s3_bucket_id
 
   # Common error - Error putting S3 notification configuration: InvalidArgument: Configuration is ambiguously defined. Cannot have overlapping suffixes in two rules if the prefixes are overlapping for the same event type.
 
   lambda_notifications = {
     lambda1 = {
-      function_arn  = module.lambda_function1.this_lambda_function_arn
-      function_name = module.lambda_function1.this_lambda_function_name
+      function_arn  = module.lambda_function1.lambda_function_arn
+      function_name = module.lambda_function1.lambda_function_name
       events        = ["s3:ObjectCreated:Put"]
       filter_prefix = "prefix/"
       filter_suffix = ".json"
     }
 
     lambda2 = {
-      function_arn  = module.lambda_function2.this_lambda_function_arn
-      function_name = module.lambda_function2.this_lambda_function_name
+      function_arn  = module.lambda_function2.lambda_function_arn
+      function_name = module.lambda_function2.lambda_function_name
       events        = ["s3:ObjectCreated:Post"]
     }
   }
@@ -137,14 +136,14 @@ module "all_notifications" {
 
   sns_notifications = {
     sns1 = {
-      topic_arn     = module.sns_topic1.this_sns_topic_arn
+      topic_arn     = module.sns_topic1.sns_topic_arn
       events        = ["s3:ObjectRemoved:Delete"]
       filter_prefix = "prefix3/"
       filter_suffix = ".csv"
     }
 
     sns2 = {
-      topic_arn = module.sns_topic2.this_sns_topic_arn
+      topic_arn = module.sns_topic2.sns_topic_arn
       events    = ["s3:ObjectRemoved:DeleteMarkerCreated"]
     }
   }