+++ examples/notification/versions.tf
@@ -1,9 +1,9 @@
 terraform {
-  required_version = ">= 0.12.6"
+  required_version = ">= 0.12.26"
 
   required_providers {
-    aws    = ">= 3.0"
-    random = ">= 2"
-    null   = ">= 2"
+    aws    = ">= 3.36"
+    random = ">= 2.0"
+    null   = ">= 2.0"
   }
 }