+++ examples/s3-replication/versions.tf
@@ -1,8 +1,8 @@
 terraform {
-  required_version = ">= 0.12.6"
+  required_version = ">= 0.12.26"
 
   required_providers {
-    aws    = ">= 3.0"
+    aws    = ">= 3.36"
     random = ">= 2.0"
   }
 }