+++ modules/notification/versions.tf
@@ -2,7 +2,6 @@ terraform {
   required_version = ">= 0.12.6"
 
   required_providers {
-    aws    = ">= 3.0"
-    random = ">= 2.0"
+    aws = ">= 3.28"
   }
 }